requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-socketio==5.1.1
python-socketio==5.3.0
eventlet==0.30.2
torch
diffusers
transformers
accelerate
safetensors
pillow
numpy
dnspython 